+++ eventsverify.go
@@ -130,7 +130,7 @@ Note: for publicly attested events, or shared protected events, you must use --t
 				leafHasher := simplehash.NewHasherV3()
 				err = leafHasher.HashEventFromV3(
 					event.V3Event,
-					simplehash.WithPrefix([]byte{uint8(LeafTypePlain)}),
+					simplehash.WithPrefix([]byte{LeafTypePlain}),
 					simplehash.WithIDCommitted(eventIDTimestamp))
 				if err != nil {
 					return err
